Fix shareButton
