+++ pubspec.yaml
@@ -68,6 +68,7 @@ dependencies:
   timeago: ^3.2.2 # ~分前などのタイムスタンプを実装する
   path: ^1.8.0 # Achievementの獲得時の画像で。cloudinaryの画像合成を利用するためにimageのURLからファイル名を取得するために利用
   cached_network_image: ^3.2.0 # ネットワーク経由で画像を表示し、画像をキャッシュディレクトリに保持する。
+  share_plus: ^4.0.4 # 連続解答記録のシェアなど
   
 
 dev_dependencies:
